requirements.txt
Django>=4.0.7, <4.1
djangorestframework>=3.13.1, <3.13.2
psycopg2>=2.9.3,<2.9.4
redis>=4.3.4, <4.4
django-oauth-toolkit>=2.1.0, <2.1.1
drf_social_oauth2>=1.2.1, <1.2.2
social-auth-app-django>=5.0.0, <5.0.1